Expect `cf push` to return non-zero status
Previously, the cloud_controller iamge contained a hack to always return
0s for all app container metrics. This commit (in tandem with rebuilding
the capi image) removes that hack. So we now no longer get any metrics
(which is at least more honest).

This change specifically modifies the smoke tests to reflect the fact
that cf push will not return a successful exit code until we have a
metrics store in place.

[#170321732] Remove the custom hack from CAPI

Co-authored-by: John Ryan <jryan@pivotal.io>