Split CI pool into two separate pools
Previously, the job that validated new commits and the job that tested
the stabilty of cf-for-k8s by running tests every 20 minutes shared the
same pool of environments.

Going forward, we'd like to grow the environment pool for testing
commits and PRs, but if the stability tests consume from the same
environment, we run into the risk of flakes consuming all available
environments. Because of this, we decided to isolate the stability
environment pool.

[#171451722]