Bump capi-k8s-release
This change updates the commit of capi-k8s-release we use in cf-for-k8s.
This bump gives us the ability now to cf push a docker-lifecycle app
without getting any errors from the CF CLI.

This is "working" because of a temporary feature to ignore errors when
log cache is not available to serve up app metrics. In the future, we
hope to return correct app metrics instead of 0's for everything.

[fixes #42]
[finishes #171480748]

Co-authored-by: Sebastian Vidrio <svidrio@pivotal.io>
Co-authored-by: Angela Chin <achin@pivotal.io>