Make ingress solution configurable
- Introduce new value called `ingress_solution_provider` which allows
operator to choose which ingress provider (i.e. Contour or Istio) they
would like to use.
- Move config/istio to config/ingress/_ytt_lib/istio
- Add ytt tests for ingress_solution_provider value
- Istio resources such as `PeerAuthentication` and `AuthorizationPolicy` now
live in the istio library and are only created if
`ingress_solution_provider` is set to Istio
- Refactor NetworkPolicy to have Istio pods selector in one place

- PS: we needed to add minio_enabled config option as a value passed into the
library. It felt better than leaving the peerauth stuff out of the
library and wrapped in an if statement.

[#175134357](https://www.pivotaltracker.com/story/show/175134357)
Co-authored-by: Kauana dos Santos <kdossantos@vmware.com>
Co-authored-by: Clay Kauzlaric <ckauzlaric@vmware.com>
Co-authored-by: Alex Standke <astandke@vmware.com>