Add netpol for eirini webhook registration job
Eirini has split the instance-index-env-injector component into a
webhook registration job and a webhook service deployment. The webhook
service requires network to and from the control plane as before, so the
eirini-instance-index-env-injector network policy is being left as is.

The new registration job requires outbound network access to the control
plane so we are adding this new network policy targeting the job.

This change is required to support versions of eirini > 2.0.0.

[#175878489]