FIX: use "skip" update-strategy for cf-k8s-networking upgrade-sidecars Job
- This job is intended to rollout all the app Pods in cf-workloads
whenever Istio is upgraded so that they can injected with newer Istio
proxy sidecars.
- Jobs in Kubernetes are meant to run once and (for the most part) have
immutable fields. If we update the cf-k8s-networking upgrade-sidecars image for any
reason (CVE fixes, changing location of the image, etc.) `kapp` will
attempt to update the `Job` if it already exists with the new image.
- Since the Job has already run and migrated the sidecars for apps we
can safely skip updating it on upgrades.

This avoids errors like:
```
kapp: Error: Applying update job/restart-workloads-for-istio1-7-3 (batch/v1) namespace: cf-workloads:
  Updating resource job/restart-workloads-for-istio1-7-3 (batch/v1) namespace: cf-workloads:
    Job.batch "restart-workloads-for-istio1-7-3" is invalid:

...
```

We have an outstanding issue on the cf-k8s-networking repo to improve
this Job to mitigate this issue in other ways:
https://github.com/cloudfoundry/cf-k8s-networking/issues/74

[#177348412](https://www.pivotaltracker.com/story/show/177348412)

Authored-by: Tim Downey <tdowney@vmware.com>