Remove image override data values
Remove image overrides for eirini, istio, minio, and postgres.
  - Remove config/values/*.yml files that set image overrides to use
  - Remove ytt overlays that overrode the images

For quarks, revert to older solution that used kbld to do overrides.
This solution is superior, because it won't apply the override if the
quarks image tag changes.
This was done by reverting (-n) 6f2122925b81c85277b3d78432b096176521dc46

[#177786619](https://www.pivotaltracker.com/story/show/177786619)

Co-authored-by: Matt Royal <mroyal@vmware.com>
Co-authored-by: Andrew Costa <ancosta@vmware.com>